feat: clean up identity action screens
