+++ src/ui/components/wallet_unlock.rs
@@ -46,7 +46,6 @@ pub trait ScreenWithWalletUnlock {
 
             // Only render the unlock prompt if the wallet requires a password and is locked
             if wallet.uses_password && !wallet.is_open() {
-                ui.add_space(20.0);
                 if let Some(alias) = &wallet.alias {
                     ui.label(format!(
                         "This wallet ({}) is locked. Please enter the password to unlock it:",
