+++ src/ui/mod.rs
@@ -11,8 +11,8 @@ use crate::ui::identities::keys::add_key_screen::AddKeyScreen;
 use crate::ui::identities::keys::key_info_screen::KeyInfoScreen;
 use crate::ui::identities::keys::keys_screen::KeysScreen;
 use crate::ui::identities::top_up_identity_screen::TopUpIdentityScreen;
-use crate::ui::identities::transfers::TransferScreen;
-use crate::ui::identities::withdraw_from_identity_screen::WithdrawalScreen;
+use crate::ui::identities::transfer_screen::TransferScreen;
+use crate::ui::identities::withdraw_screen::WithdrawalScreen;
 use crate::ui::network_chooser_screen::NetworkChooserScreen;
 use crate::ui::tools::proof_log_screen::ProofLogScreen;
 use crate::ui::wallets::import_wallet_screen::ImportWalletScreen;
